[Android]: Fix incorrect use of jint for pointer in ContentViewCoreImpl
BUG=354405,346626

Review URL: https://codereview.chromium.org/240003009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@264604 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3cb3ff0a4668800acd64bbf91a237c3344622f0f

+++ content/browser/android/content_view_core_impl.cc
@@ -1377,7 +1377,7 @@ void ContentViewCoreImpl::WasResized(JNIEnv* env, jobject obj) {
 }
 
 void ContentViewCoreImpl::ShowInterstitialPage(
-    JNIEnv* env, jobject obj, jstring jurl, jint delegate_ptr) {
+    JNIEnv* env, jobject obj, jstring jurl, jlong delegate_ptr) {
   GURL url(base::android::ConvertJavaStringToUTF8(env, jurl));
   InterstitialPageDelegateAndroid* delegate =
       reinterpret_cast<InterstitialPageDelegateAndroid*>(delegate_ptr);
